feat: hash password before save
