+++ src/auth/auth.service.ts
@@ -1,11 +1,16 @@
 import { Injectable } from "@nestjs/common";
+import { DbService } from "@app/db/db.service";
+import { CreateUserDto } from "@app/user/dto/create-user.dto";
+
 
 @Injectable()
 export class AuthService {
 
+  constructor(private db: DbService) {
+  }
 
-  async signup() {
-    //
+  async signup(dto: CreateUserDto) {
+    return this.db.user.create(dto);
   }
